axel-server/pyproject.toml
[project]
name = "axel-server"
version = "0.1.0"
description = "Servidor IoT para monitoreo de luminosidad con Arduino y MQTT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pyserial>=3.5",
    "rich>=14.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"